restructure code: split reading input from processing commands. Refactor .h
