+++ include/list.h
@@ -1,22 +1,5 @@
-#ifndef INCLUDE_H
-#define INCLUDE_H
-
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
-#include <unistd.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <fcntl.h>
-#include <ctype.h>
-
-#include "option.h"
-
-#endif
-
 #ifndef LIST_H
 #define LIST_H
-
 #define SELLECT_ALL "*"
 
 typedef struct node_s {
@@ -25,7 +8,7 @@ typedef struct node_s {
     struct node_s *next; 
 }node_t;
 
-node_t *init_list(char *bid, int nid);
+node_t *init_list(void);
 node_t *add_node(node_t *head, char *bid, int nid);
 node_t *add_block(node_t *head, char *bid, int nid);
 node_t *remove_node(node_t *head, int nid);
@@ -40,6 +23,4 @@ void print_block_list(node_t *head, char *bid);
 
 void free_list(node_t *head);
 
-
-option_t process_input(int std_in, node_t *data);
 #endif
