+++ src/my_blockchain.c
@@ -18,14 +18,22 @@
 
 #include <stdlib.h>
 #include  <stdio.h>
-#include "../include/list.h"
+#include "../include/my_blockchain.h"
+#define MAX_BUFF_SIZE 100
 
 int main(void)
 {
-    node_t *unsynced = NULL;
-    while(process_input(STDIN_FILENO, unsynced) != QUIT){
+    node_t *unsynced = init_list();
+    char *input_buffer = malloc(sizeof(char)*MAX_BUFF_SIZE);
+
+    while(process_input(STDIN_FILENO, input_buffer) != QUIT){
+
 
     } 
+
+
+
+
     /* printf("cmd: %s, type: %s, bid: %s, nid: %i\n", input->cmd, input->type, input->bid, input->nid); */
 
     /* TODO: NEED FUNCTION TO FREE THESE */
